rename authentication and move get_current_user method
