+++ src/aiod/__init__.py
@@ -1,5 +1,4 @@
-from aiod.authorisation.authorisation import login, logout
-from aiod.default.get_current_user import get_current_user
+from aiod.authentication.authentication import login, logout, get_current_user
 from aiod.resources import case_studies
 from aiod.resources import computational_assets
 from aiod.resources import contacts
